+++ code/classeine-lib/clsn/ui/renderers/content_pane_renderer.cpp
@@ -3,32 +3,30 @@
 // SPDX-License-Identifier: BSD-3-Clause
 // SPDX-FileCopyrightText: © 2024 Ernesto Bascón Pantoja
 
-#include "dynamic_pane_renderer.h"
+#include "content_pane_renderer.h"
 
-#include "clsn/ui/dynamic_pane.h"
+#include "clsn/ui/dynamic_content_pane.h"
 
 #include "clsn/ui/painters/background_painter.h"
-
-#include <clsn/ui/ui_manager.h>
 
 namespace clsn::ui::renderers
 {
     using namespace clsn::draw;
 
-    void dynamic_pane_renderer::paint(graphics& gfx,
+    void content_pane_renderer::paint(graphics& gfx,
                               const region& rgn,
                               const control& ctrl) const
     {
-        const auto& dyn_pane = static_cast<const dynamic_pane&>(ctrl);
+        const auto& con_pane = static_cast<const content_pane&>(ctrl);
 
-        const auto& inner_ctrl = dyn_pane.get_inner_control();
-        if (inner_ctrl == nullptr)
+        auto inner_ctrl = con_pane.get_actual_inner_control();
+        if (!inner_ctrl.has_value())
         {
-            const auto& clr = dyn_pane.get_actual_background_color();
+            const auto& clr = con_pane.get_actual_background_color();
             painters::background_painter::paint_background(gfx, rgn, clr);
             return;
         }
 
-        inner_ctrl->paint(gfx, rgn);
+        inner_ctrl.get_ref().paint(gfx, rgn);
     }
 }