+++ code/classeine-lib/clsn/ui/window.cpp
@@ -9,23 +9,6 @@
 
 #include <clsn/core/console.h>
 
-namespace
-{
-    using namespace clsn::ui;
-
-    auto are_equal(
-            const const_optional_reference<control>& a,
-            const const_optional_reference<control>& b) -> bool
-    {
-        if (a.has_value() != b.has_value())
-            return false;
-
-        if (!a.has_value())
-            return true;
-
-        return a.value().get() == b.value().get();
-    }
-}
 namespace clsn::ui
 {
     window::window(std::string_view section_name)
@@ -57,14 +40,14 @@ namespace clsn::ui
 
     void window::grab_mouse(control& ctrl)
     {
-        m_mouse_grabber_control = std::ref(ctrl);
+        m_mouse_grabber_control = ctrl;
     }
 
     void window::release_mouse()
     {
         if (m_mouse_grabber_control.has_value())
         {
-            m_mouse_grabber_control.value().get().release_mouse();
+            m_mouse_grabber_control.get_ref().release_mouse();
             m_mouse_grabber_control = std::nullopt;
         }
     }
@@ -72,19 +55,19 @@ namespace clsn::ui
     void window::process_mouse_moved_event(events::mouse_moved_event &e)
     {
         auto currently_hovered = get_control_by_position(e.get_position());
-        if (!are_equal(m_hovered_control, currently_hovered))
+        if (m_hovered_control != currently_hovered)
         {
             if (m_hovered_control.has_value())
             {
-                m_hovered_control.value().get().invalidate();
+                m_hovered_control.get_ref().invalidate();
             }
 
             m_hovered_control = currently_hovered;
 
             if (m_hovered_control.has_value())
             {
-                clsn::core::console::debug("window::process_mouse_moved_event: Ctrl: {}", m_hovered_control.value().get());
-                m_hovered_control.value().get().invalidate();
+                clsn::core::console::debug("window::process_mouse_moved_event: Ctrl: {}", m_hovered_control.get_ref());
+                m_hovered_control.get_ref().invalidate();
             }
         }
     }
@@ -94,7 +77,7 @@ namespace clsn::ui
         if (!m_hovered_control.has_value())
             return false;
 
-        return control == m_hovered_control.value().get();
+        return control == m_hovered_control.get_ref();
     }
 
 }